typo + default zshrc

+++ scripts/setup.sh
@@ -22,7 +22,7 @@ install_node() {
     echo -e "\033[32m✔ Node is installed \e[0m"
   else
     echo -e "\033[31m❌ Node missing\e[0m"
-    echo -e "\033[36Installing now...\e[0m"
+    echo -e "\033[36mInstalling now...\e[0m"
     curl -sL https://deb.nodesource.com/setup_14.x | sudo bash -
     sudo apt install -y nodejs
     # seperate because npm needs nodejs before it can install
@@ -38,7 +38,7 @@ install_rmate() {
     echo -e "\033[32m✔ rmate is installed\e[0m"
   else
     echo -e "\033[31m❌ rmate missing\e[0m"
-    echo -e "\033[36Installing now...\e[0m"
+    echo -e "\033[36mInstalling now...\e[0m"
     sudo wget -O /usr/local/bin/rsub \https://raw.github.com/aurora/rmate/master/rmate
     sudo chmod a+x /usr/local/bin/rsub
   fi
@@ -66,6 +66,13 @@ install_python() {
   fi
 }
 
+download_zshrc() {
+  #download .zshrc from github
+  wget https://raw.githubusercontent.com/WDaan/dotfiles/master/.zshrc -O /home/$username/.zshrc
+  sed -i "s/YOUR_USERNAME/$username/g" /home/$username/.zshrc
+  chown $username /home/$username/.zshrc
+}
+
 configure_zsh() {
   if which zsh >/dev/null; then
     echo -e "\033[32m✔ Zsh is installed\e[0m"
@@ -84,28 +91,29 @@ configure_zsh() {
     git clone https://github.com/zsh-users/zsh-autosuggestions ${ZSH_CUSTOM:-/home/$username/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
   fi
 
-  #restore .zshrc
-  echo -e "\033[36mrestore .zshrc?\e[0m"
-  PS3='Please enter your choice: '
-  options=("Yes" "No")
-  select opt in "${options[@]}"; do
-    case $opt in
-    "Yes")
-      rm /home/$username/.zshrc
-      cd ~
-      #download .zshrc from github
-      wget https://raw.githubusercontent.com/WDaan/dotfiles/master/.zshrc -O /home/$username/.zshrc
-      sed -i "s/YOUR_USERNAME/$username/g" /home/$username/.zshrc
-      chown $username /home/$username/.zshrc
-      echo -e "\033[36m✔ Restored .zshrc to default\e[0m"
-      break
-      ;;
-    *)
-      echo -e "\033[36mOK than...\e[0m"
-      break
-      ;;
-    esac
-  done
+  if [ -f /home/$username/.zshrc ]; then
+    #restore .zshrc
+    echo -e "\033[36mrestore .zshrc?\e[0m"
+    PS3='Please enter your choice: '
+    options=("Yes" "No")
+    select opt in "${options[@]}"; do
+      case $opt in
+      "Yes")
+        rm /home/$username/.zshrc
+        cd ~
+        download_zshrc
+        echo -e "\033[36m✔ Restored .zshrc to default\e[0m"
+        break
+        ;;
+      *)
+        echo -e "\033[36mOK than...\e[0m"
+        break
+        ;;
+      esac
+    done
+  else
+    download_zshrc
+  fi
 }
 
 install_kubectl() {
